requirements.txt
django==3.0.7
django-filter==2.3.0
djangorestframework==3.11.0